graphcolourer
