+++ colouring.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <time.h>
 #include <string.h>
+#include <stdlib.h>
 #include "node.h"
 #include "graphgenerator.h"
 #include "graphcolourer.h"
